po: do not hardcode the path to msgfmt
It might be found in some other path, so just let it use the system's
configured path.

For example, when building with yocto/bitbake the build system will set
up msgfmt in their native-tools path and this package will fail to build
because it looks for msgfmt on /usr/bin (looking at the host system) and
it might not be available there.

Change-Id: I95769aab97600d366f6d2c1988f4ae3ab9f80817